update: refactoring works now. challenges: I had to make every folder as a module by adding __init__.py. But pycharm still shows warning. I fixed this by marking every packages as sources root.
I was quite amazed that how it works and runs. LOL.
