+++ .idea/april-music-player-os.iml
@@ -2,9 +2,17 @@
 <module type="PYTHON_MODULE" version="4">
   <component name="NewModuleRootManager">
     <content url="file://$MODULE_DIR$">
+      <sourceFolder url="file://$MODULE_DIR$/april-music-player" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/april-music-player/assets" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/april-music-player/components" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/april-music-player/consts" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/april-music-player/main_ui" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/april-music-player/music_player" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/april-music-player/plugins" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/april-music-player/_utils" isTestSource="false" />
       <excludeFolder url="file://$MODULE_DIR$/.venv" />
     </content>
-    <orderEntry type="jdk" jdkName="Python 3.13 (april-music-player-os)" jdkType="Python SDK" />
+    <orderEntry type="jdk" jdkName="Python 3.13" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
   <component name="PyDocumentationSettings">
