+++ april-music-player/components/addnewdirectory.py
@@ -3,7 +3,7 @@ from PyQt6.QtWidgets import (
     QCheckBox, QScrollArea, QWidget, QMessageBox
 )
 from PyQt6.QtCore import Qt
-from utils.easy_json import EasyJson
+from _utils.easy_json import EasyJson
 
 
 class AddNewDirectory(QDialog):
